Big revamp of yellowfin integration
Command-line changes:
Setpassword can now change a yellowfin password. This is used
by our installer to change the 'admin' password for yellowfin.

Createuser can now create a user in yellowfin. I am going to remove
this in the next commit. See that commit for details.

HTTP API changes:
Change the yellowfin 'admin' user to 'admin@yellowfin.com.au', as
this is the user that the yellowfin installer creates.

Upon login via HTTP, create a yellowfin user if one does not already exist.
Then, update the yellowfin role to correspond with the IMQS roles.
If the IMQS roles list has no yellowfin boxes ticked, then do not do
any of this (ie do not login to yellowfin, nor create user, etc).

Fix some code paths in the yellowfin requests that could result
in false positives, when in fact the call to yellowfin actually failed.