Reverted createuser from affecting Yellowfin
The previous commit changed the command-line "createuser"
command so that it created a user in both Authaus and Yellowfin.
Our policy, however, is to create the yellowfin user automatically
during HTTP login. For consistency sake, I have removed the
yellowfin functionality of the command-line "createuser" tool.

Also, the command-line "showidentities" command now
behaves better in the face of there being no permit for
a particular user.