Add -nosvc option, which disables Windows Service launch
When running tests under Jenkins, we need a way to launch imqsauth without
it thinking that it is a service. The way it detects whether it is a
service, is by checking if the 'session' is interactive. Now, since
Jenkins is running as a service, and imqsauth is a child process of that
service, imqsauth will not be running in an interactive session - and so
it believes it must run as a service.
The -nosvc command-line options allows us to disable the automatic
service detection logic.