Edited Dockerfile and go.mod for the sake of Docker Builds
The imqsauth builds were not working, for a few reasons. I happened upon
this while I needed to make new permissions for the Leasing module. In
order for this commit to be made, I needed to release new versions of
the following other repos:
- IMQS/authaus
- IMQS/serviceauth
- IMQS/log

Now the Docker builds for auth can hopefully be built, or at least is in
a state where it can be built.