PM-1363: Fix slow fetch for user management endpoints
Relates to /userobjects call from the front-end.

One culprit is many calls to authaus.GroupIDsToNames. The new version
of this function takes an additional 'cache' parameter which reduces
the number of DB queries.

The other issue was a loop with db accesses for each user when building user display names. The existing user list already at hand was transformed to a map for re-use in the JSON object build code.