Ensure group changes works on user with no permit (#101)
Ensure group changes works on user with no permit. In cases where a user has been newly created there is no permit yet. To cater for this a special check against the ErrIdentityPermitNotFound error has been added. That will then ignore this error since that means there is not permit that could be found.

refs: NEXUS-4289